// apps/admin/services/common-api/types.gen.ts
export type AuthConfig = {
    device: DeviceAuthticateConfig;
    email: EmailAuthticateConfig;
    mobile: MobileAuthenticateConfig;
    register: PubilcRegisterConfig;
};

export type CheckVerificationCodeRequest = {
    account: string;
    code: string;
    method: string;
    type: number;
};

export type CheckVerificationCodeRespone = {
    status: boolean;
};

export type Currency = {
    currency_symbol: string;
    currency_unit: string;
};

export type DeviceAuthticateConfig = {
    enable: boolean;
    enable_security: boolean;
    only_real_device: boolean;
    show_ads: boolean;
};

export type DownloadLink = {
    android?: string;
    harmony?: string;
    ios?: string;
    linux?: string;
    mac?: string;
    windows?: string;
};

export type EmailAuthticateConfig = {
    domain_suffix_list: string;
    enable: boolean;
    enable_domain_suffix: boolean;
    enable_verify: boolean;
};
// ...
export type GetGlobalConfigResponse = {
    auth: AuthConfig;
    currency: Currency;
    invite: InviteConfig;
    oauth_methods: Array<string> | null;
    site: SiteConfig;
    subscribe: SubscribeConfig;
    verify: VeifyConfig;
    verify_code: PubilcVerifyCodeConfig;
    web_ad: boolean;
};
// ...
export type InviteConfig = {
    forced_invite: boolean;
    only_first_purchase: boolean;
    referral_percentage: number;
};

export type MobileAuthenticateConfig = {
    enable: boolean;
    enable_whitelist: boolean;
    whitelist: Array<string> | null;
};

export type PrivacyPolicyConfig = {
    privacy_policy: string;
};

export type PubilcRegisterConfig = {
    enable_ip_register_limit: boolean;
    ip_register_limit: number;
    ip_register_limit_duration: number;
    stop_register: boolean;
};

export type PubilcVerifyCodeConfig = {
    verify_code_interval: number;
};

export type SendCodeRequest = {
    email: string;
    type: number;
};

export type SendCodeResponse = {
    code?: string;
    status: boolean;
};

export type SendSmsCodeRequest = {
    telephone: string;
    telephone_area_code: string;
    type: number;
};

export type SiteConfig = {
    custom_data: string;
    custom_html: string;
    host: string;
    keywords: string;
    site_desc: string;
    site_logo: string;
    site_name: string;
};

export type SubscribeClient = {
    description?: string;
    download_link?: DownloadLink;
    icon?: string;
    id: number;
    is_default: boolean;
    name: string;
    scheme?: string;
};

export type SubscribeConfig = {
    pan_domain: boolean;
    single_model: boolean;
    subscribe_domain: string;
    subscribe_path: string;
    user_agent_limit: boolean;
    user_agent_list: string;
};

export type VeifyConfig = {
    enable_login_verify: boolean;
    enable_register_verify: boolean;
    enable_reset_password_verify: boolean;
    turnstile_site_key: string;
};